Refactor commands, simplify registration, add QR scanning
Refactored command names and properties for clarity and consistency.
Removed email verification logic, focusing on phone verification.
Added QR code scanning functionality for onboarding invitations.
Simplified account creation logic, removing alternative name generation.
Updated UI elements in ContactSupportView.xaml and NameEntryView.xaml.
Enhanced validation and navigation logic in view models.
